Move the grouping code to drizzle
